// frontend/sdk/model/referenceItem.ts
/**
 * Generated by orval v8.5.2 🍺
 * Do not edit manually.
 * Shipping Disruption Management API
 * MVP v1 REST API for ACME Inc. to ingest shipping disruption signals, enrich relevance and impact, escalate incidents, and generate action proposals for manager decisioning.
 * OpenAPI spec version: 1.0.0
 */
import type { ReferenceItemMetadata } from './referenceItemMetadata';

export interface ReferenceItem {
  id: string;
  name: string;
  /** @nullable */
  code?: string | null;
  metadata?: ReferenceItemMetadata;
}